Create a dotfiles repository
The purpose of this repository is to describe my personal customizations
for software I use on my workstation.

The customizations are applied by using a configuration management
framework called salt and described using salt states in the form of sls
files. In short, my experience with salt so far has been positive. There
are still a few TODOs to complete, but this commit captures a good
portion of my existing setup.

The lack of history is mostly due to squashing a bunch of
work-in-progress commits to start the history off a little bit cleaner
as I was learning salt.