Automate the Gentoo install
Create and provision a Gentoo machine image using packer in a virtualbox
environment.

Packer is used to boot the offical Gentoo minimal installation ISO and
provision a new Gentoo VM. The provisioning is done using shell scripts
to execute the steps documented in the Gentoo installation handbook.

The output VM is intended to be used in additional provisioning steps
and some best security practices are not applied at this point (e.g.,
simple root password, sshd permit root login, etc.)

The lack of history is due to squashing work-in-progress commits to
start the history off a little bit cleaner as I was learning packer.