Generate initramfs file list
The initramfs file list processed by the kernel build to create the
initramfs is now generated by a script. The notable change here is the
use of lddtree to output the file paths for dependencies of the commands
used in the initramfs init script. This helps to greatly reduce
maintenance needed to the initramfs file list when the dependencies for
the init commands are updated due to the version being included in the
file paths.