Encrypt using dm-crypt with LUKS
Encrypt root and swap volumes using dm-crypt with LUKS. The setup
demonstrates that the initramfs works with an encrypted root, but
currently uses some insecure practices such as using urandom to generate
the volume key and stores the key file in the initramfs in order to ease
the second stage provisioning using salt. This will be addressed in a
later commit.