Configure Gentoo mirrors in make.conf
Gentoo mirrors are now configured in make.conf to be used in emerge, not
just as a release mirror. This will also allow for configuration of
multiple mirrors.

Added GENTOO_MIRRORS to .env, along with RELEASE_MIRROR to continue to
allow for separation between package and release mirrors if desired,
otherwise falls back to the first mirror in GENTOO_MIRRORS.