Remove VirtualBox builder
Simplify the build process and reduce maintenance by removing the Packer
VirtualBox builder and guest support. I'm not currently planning on
using this now that the Qemu builder is default. Although, it may have
been nice to keep the VirtualBox support in place in case this needed to
be used outside of Linux, if that becomes important this can be brought
back in.