Export autobuild/vm variables to environment
Export the autobuild and vm variables to the environment so that they
can be easily used in future scripts.

Adapt the packer templates to explicitly define all user variables,
packer treats user variables as the single source of input to a template
and can only use environment variables through the default values of
user variables. Make the exported variables the default values.