Make kernel config choices obvious
The current Linux kernel configuration is a generated file that has not
been maintained and at nearly 4000 lines it is hard to reason about the
selected features and parameters.

Fortunatly the kernel provides a tool to help manipulate the config
options (scripts/config). Remove the old config file and replace it with
the default config options provided by the kernel maintainers using the
defconfig make target, add to those options using the config tool.

This approach to the kernel config will hopefully help to keep it
maintainable and better documented.