Use less strict package version specifiers
Use version specifier `~` in place of `=` for passff-host as this
package revision is bumped more often than the version, usually for
general ebuild maintenance in this case to add Python 3.8 and 3.9
compatibility.

Match all zeromq versions for USE flag.

The nitrokey app is stable so there's no need to accept keywords.