Use the bios option to workaround a Packer issue
The qemuargs are supposed to override matching default qemu switch/value
pairs which presents a conflict for drive arguments since these are
allowed to be used more than once. The issue in this case is the drive
added to boot to the TianoCore EDK2 UEFI for QEMU (OVMF) overrides the
disk image drive Packer creates so the UEFI boots with no operating
system. There doesn't appear to be a way to specify additional drives
while using the disk image config and attempts to simplify in Packer
have been attempted: https://github.com/hashicorp/packer/issues/8436.

Although the drive option is probably the preferred option for OVMF, the
bios option is also supported and seems to be the simplest alternative
to workaround the Packer issue.