Replace Makefile with script
The Makefile acts basically as a script where the major benifits of
using make are not realized. The rules used are currently very simple
because most of the actual build complexity is already contained in the
programs that are launched by the Makefile. The intention is to replace
the Makefile with a script so that it can be used for general purpose
commands and provide a more unified interface between build tools (e.g.
Packer, Salt).