Simplify boot command
The Gentoo minimal install iso provides options on boot to set the root
password and start the SSH daemon. Using this allows the boot command to
be simplified to just a single command and helps to avoid timing issues
for the command sequence that sometimes occured if the system was under
load.