Upgrade Packer to use HCL2 template files
There wasn't a real need to use HCL2 over JSON, but being stuck at home
what would be better to do than to try it out?

Using HCL2 allows for up front type checking for variables rather than
indirectly once they were used, actual comments, etc. It seems to be the
way forward for Packer and does allow for things that otherwise wouldn't
be supported in JSON templates, so it could be useful in the future.

HCL2 is a beta feature of Packer since 1.5.0, the Gentoo ebuild
repository doesn't include this version yet but it seems to be good
enough to bump the ebuild version for now (Example in README). In Packer
1.6.0 the `iso_checksum_type` fields have been removed and the type can
now be inferred, this change also removes this field so at least a 1.6.0
version is required.