Exclude .saltignore from copy in salt-top
The .saltignore file can list patterns for files to ignore, this change
will now implicitly include the .saltignore file to be ignored too.

When using dotglob with extglob to exclude .saltignore like
`!(.saltignore)` the '.' and '..' paths were not excluded. This seems to
be the case when using '.' at the start to cause an explicit match for
'.' and '..'. Using `[.]` in `!([.]saltignore)` works around this, the
pattern `!(.|..|.saltignore)` could have also been used.

I stumbled on this work around and it wasn't clear until I found someone
else trying something similar on StackExchange[1].

[1]: https://unix.stackexchange.com/a/382124