Make use of SOURCE consistent
The SOURCE variable is used to provide additional flexibility to offset
the relative path of the configuration when staged to some temporary
location in the VM or host machine.

This change makes the SOURCE base path consistent between the iso and
salt builds starting at the perspective of the project root, as is
currently the case for all scripts in the project.