Remove the minion id config
The minion id will be generated if not specified, this is fine for a
standalone use case as we don't need to match by id at the moment. The
id will still be assigned, but salt will take some steps to find a good
value trying the fully-qualified domain name of the machine, falling
back to a routable IP address, then finally 'localhost'.