Place generated files into build specific output directory
Generated files are now placed into the build specific output directory
so they may be cleaned if the build is interrupted. This is done by
exposing the output_directory and iso_url configuration and avoiding the
need to generate Packer HCL variable files for both iso and salt Gentoo
VM build types.